refactor(rolldown_plugin_transform): move utils logic to utils.rs
